requirements.txt
asgiref==3.3.1
Django==3.1.6
django-debug-toolbar==3.2
django-dotenv==1.4.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
Pillow==8.1.0
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
drf-nested-routers==0.92.5
